backend/apps/intelligence-ms/requirements.txt
grpcio>=1.76.0
grpcio-tools>=1.76.0
pymongo>=4.9.0
python-dotenv>=1.0.1
nltk>=3.9.0
numpy>=2.1.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
transformers>=4.42.0
torch>=2.3.0
